Renamed genedatatypeagentnew.ts to genedatatypenew.ts
