+++ server/routes/chat/genedatatypenew.ts
