+++ server/routes/chat/phrase2entity.ts
@@ -1,6 +1,6 @@
 import type { LlmConfig, GeneDataTypeResult, GeneSetDataTypeResult } from '#types'
 import { extractGenesetsFromPromptNew, extractGenesFromPrompt, getGenesetNames } from './utils.ts'
-import { classifyGeneDataTypePhrase } from './genedatatypeagentnew.ts'
+import { classifyGeneDataTypePhrase } from './genedatatypenew.ts'
 import { classifyGeneSetDataType, GENE_SET_KEYWORDS } from './genesetdatatypeagent.ts'
 import { GENE_FEATURE_KEYWORDS, determineAmbiguousGenePrompt } from './ambiguousgeneagent.ts'
 import { getDsAllowedTermTypes } from '../termdb.config.ts'
